+++ source/databricks/tests/integration/calculator/steps/test_adjust_production.py
@@ -19,7 +19,7 @@ from geh_stream.codelists import (
     ResolutionDuration,
     MarketEvaluationPointType,
 )
-from geh_stream.aggregation_utils.aggregators import adjust_flex_consumption
+from package.steps import adjust_production
 from geh_stream.codelists import Quality
 from geh_stream.shared.data_classes import Metadata
 from geh_stream.aggregation_utils.aggregation_result_formatter import (
@@ -41,10 +41,10 @@ default_domain = "D1"
 default_responsible = "R1"
 default_supplier = "S1"
 default_sum_quantity = Decimal(1)
-default_added_grid_loss = Decimal(3)
+default_added_system_correction = Decimal(3)
 default_aggregated_quality = Quality.estimated.value
 default_resolution = ResolutionDuration.hour
-default_metering_point_type = MarketEvaluationPointType.consumption.value
+default_metering_point_type = MarketEvaluationPointType.production.value
 
 date_time_formatting_string = "%Y-%m-%dT%H:%M:%S%z"
 default_time_window = {
@@ -58,11 +58,13 @@ default_valid_to = datetime.strptime(
     "2020-01-01T01:00:00+0000", date_time_formatting_string
 )
 
+metadata = Metadata("1", "1", "1", "1", "1")
+
 
 @pytest.fixture(scope="module")
-def flex_consumption_result_schema():
+def hourly_production_result_schema():
     """
-    Input flex consumption result data frame schema
+    Input hourly production result data frame schema
     """
     return (
         StructType()
@@ -84,14 +86,14 @@ def flex_consumption_result_schema():
 
 
 @pytest.fixture(scope="module")
-def added_grid_loss_result_schema():
+def added_system_correction_result_schema():
     """
-    Input grid loss result schema
+    Input system correction result schema
     """
     return (
         StructType()
         .add(Colname.grid_area, StringType(), False)
-        .add(Colname.added_grid_loss, DecimalType())
+        .add(Colname.added_system_correction, DecimalType())
         .add(
             Colname.time_window,
             StructType()
@@ -107,9 +109,9 @@ def added_grid_loss_result_schema():
 
 
 @pytest.fixture(scope="module")
-def grid_loss_sys_cor_schema():
+def sys_cor_schema():
     """
-    Input grid loss system correction data frame schema
+    Input system correction data frame schema
     """
     return (
         StructType()
@@ -118,7 +120,7 @@ def grid_loss_sys_cor_schema():
         .add(Colname.energy_supplier_id, StringType())
         .add(Colname.from_date, TimestampType())
         .add(Colname.to_date, TimestampType())
-        .add(Colname.is_grid_loss, BooleanType())
+        .add(Colname.is_system_correction, BooleanType())
     )
 
 
@@ -150,7 +152,7 @@ def expected_schema():
 
 
 @pytest.fixture(scope="module")
-def flex_consumption_result_row_factory(spark, flex_consumption_result_schema):
+def hourly_production_result_row_factory(spark, hourly_production_result_schema):
     """
     Factory to generate a single row of  data, with default parameters as specified above.
     """
@@ -177,20 +179,22 @@ def flex_consumption_result_row_factory(spark, flex_consumption_result_schema):
                 Colname.metering_point_type: [metering_point_type],
             }
         )
-        return spark.createDataFrame(pandas_df, schema=flex_consumption_result_schema)
+        return spark.createDataFrame(pandas_df, schema=hourly_production_result_schema)
 
     return factory
 
 
 @pytest.fixture(scope="module")
-def added_grid_loss_result_row_factory(spark, added_grid_loss_result_schema):
+def added_system_correction_result_row_factory(
+    spark, added_system_correction_result_schema
+):
     """
     Factory to generate a single row of  data, with default parameters as specified above.
     """
 
     def factory(
         domain=default_domain,
-        added_grid_loss=default_added_grid_loss,
+        added_system_correction=default_added_system_correction,
         time_window=default_time_window,
         sum_quantity=default_sum_quantity,
         aggregated_quality=default_aggregated_quality,
@@ -200,7 +204,7 @@ def added_grid_loss_result_row_factory(spark, added_grid_loss_result_schema):
         pandas_df = pd.DataFrame(
             {
                 Colname.grid_area: [domain],
-                Colname.added_grid_loss: [added_grid_loss],
+                Colname.added_system_correction: [added_system_correction],
                 Colname.time_window: [time_window],
                 Colname.sum_quantity: [sum_quantity],
                 Colname.quality: [aggregated_quality],
@@ -208,13 +212,15 @@ def added_grid_loss_result_row_factory(spark, added_grid_loss_result_schema):
                 Colname.metering_point_type: [metering_point_type],
             }
         )
-        return spark.createDataFrame(pandas_df, schema=added_grid_loss_result_schema)
+        return spark.createDataFrame(
+            pandas_df, schema=added_system_correction_result_schema
+        )
 
     return factory
 
 
 @pytest.fixture(scope="module")
-def grid_loss_sys_cor_row_factory(spark, grid_loss_sys_cor_schema):
+def sys_cor_row_factory(spark, sys_cor_schema):
     """
     Factory to generate a single row of  data, with default parameters as specified above.
     """
@@ -225,7 +231,7 @@ def grid_loss_sys_cor_row_factory(spark, grid_loss_sys_cor_schema):
         supplier=default_supplier,
         valid_from=default_valid_from,
         valid_to=default_valid_to,
-        is_grid_loss=True,
+        is_system_correction=True,
     ):
         pandas_df = pd.DataFrame(
             {
@@ -234,72 +240,69 @@ def grid_loss_sys_cor_row_factory(spark, grid_loss_sys_cor_schema):
                 Colname.energy_supplier_id: [supplier],
                 Colname.from_date: [valid_from],
                 Colname.to_date: [valid_to],
-                Colname.is_grid_loss: [is_grid_loss],
+                Colname.is_system_correction: [is_system_correction],
             }
         )
-        return spark.createDataFrame(pandas_df, schema=grid_loss_sys_cor_schema)
+        return spark.createDataFrame(pandas_df, schema=sys_cor_schema)
 
     return factory
 
 
-metadata = Metadata("1", "1", "1", "1", "1")
-
-
-def test_grid_area_grid_loss_is_added_to_grid_loss_energy_responsible(
-    flex_consumption_result_row_factory,
-    added_grid_loss_result_row_factory,
-    grid_loss_sys_cor_row_factory,
+def test_grid_area_system_correction_is_added_to_system_correction_energy_responsible(
+    hourly_production_result_row_factory,
+    added_system_correction_result_row_factory,
+    sys_cor_row_factory,
 ):
     results = {}
     results[
-        ResultKeyName.flex_consumption
+        ResultKeyName.hourly_production
     ] = create_dataframe_from_aggregation_result_schema(
-        metadata, flex_consumption_result_row_factory(supplier="A")
+        metadata, hourly_production_result_row_factory(supplier="A")
     )
 
     results[
-        ResultKeyName.added_grid_loss
+        ResultKeyName.added_system_correction
     ] = create_dataframe_from_aggregation_result_schema(
-        metadata, added_grid_loss_result_row_factory()
+        metadata, added_system_correction_result_row_factory()
     )
 
-    results[
-        ResultKeyName.grid_loss_sys_cor_master_data
-    ] = grid_loss_sys_cor_row_factory(supplier="A")
+    results[ResultKeyName.grid_loss_sys_cor_master_data] = sys_cor_row_factory(
+        supplier="A"
+    )
 
-    result_df = adjust_flex_consumption(results, metadata)
+    result_df = adjust_production(results, metadata)
 
     assert (
         result_df.filter(col(Colname.energy_supplier_id) == "A").collect()[0][
             Colname.sum_quantity
         ]
-        == default_added_grid_loss + default_sum_quantity
+        == default_added_system_correction + default_sum_quantity
     )
 
 
 def test_grid_area_grid_loss_is_not_added_to_non_grid_loss_energy_responsible(
-    flex_consumption_result_row_factory,
-    added_grid_loss_result_row_factory,
-    grid_loss_sys_cor_row_factory,
+    hourly_production_result_row_factory,
+    added_system_correction_result_row_factory,
+    sys_cor_row_factory,
 ):
     results = {}
     results[
-        ResultKeyName.flex_consumption
+        ResultKeyName.hourly_production
     ] = create_dataframe_from_aggregation_result_schema(
-        metadata, flex_consumption_result_row_factory(supplier="A")
+        metadata, hourly_production_result_row_factory(supplier="A")
     )
 
     results[
-        ResultKeyName.added_grid_loss
+        ResultKeyName.added_system_correction
     ] = create_dataframe_from_aggregation_result_schema(
-        metadata, added_grid_loss_result_row_factory()
+        metadata, added_system_correction_result_row_factory()
     )
 
-    results[
-        ResultKeyName.grid_loss_sys_cor_master_data
-    ] = grid_loss_sys_cor_row_factory(supplier="B")
+    results[ResultKeyName.grid_loss_sys_cor_master_data] = sys_cor_row_factory(
+        supplier="B"
+    )
 
-    result_df = adjust_flex_consumption(results, metadata)
+    result_df = adjust_production(results, metadata)
 
     assert (
         result_df.filter(col(Colname.energy_supplier_id) == "A").collect()[0][
@@ -310,32 +313,32 @@ def test_grid_area_grid_loss_is_not_added_to_non_grid_loss_energy_responsible(
 
 
 def test_result_dataframe_contains_same_number_of_results_with_same_energy_suppliers_as_flex_consumption_result_dataframe(
-    flex_consumption_result_row_factory,
-    added_grid_loss_result_row_factory,
-    grid_loss_sys_cor_row_factory,
+    hourly_production_result_row_factory,
+    added_system_correction_result_row_factory,
+    sys_cor_row_factory,
 ):
     results = {}
-    fc_row_1 = flex_consumption_result_row_factory(supplier="A")
-    fc_row_2 = flex_consumption_result_row_factory(supplier="B")
-    fc_row_3 = flex_consumption_result_row_factory(supplier="C")
+    hp_row_1 = hourly_production_result_row_factory(supplier="A")
+    hp_row_2 = hourly_production_result_row_factory(supplier="B")
+    hp_row_3 = hourly_production_result_row_factory(supplier="C")
 
     results[
-        ResultKeyName.flex_consumption
+        ResultKeyName.hourly_production
     ] = create_dataframe_from_aggregation_result_schema(
-        metadata, fc_row_1.union(fc_row_2).union(fc_row_3)
+        metadata, hp_row_1.union(hp_row_2).union(hp_row_3)
     )
 
     results[
-        ResultKeyName.added_grid_loss
+        ResultKeyName.added_system_correction
     ] = create_dataframe_from_aggregation_result_schema(
-        metadata, added_grid_loss_result_row_factory()
+        metadata, added_system_correction_result_row_factory()
     )
 
-    results[
-        ResultKeyName.grid_loss_sys_cor_master_data
-    ] = grid_loss_sys_cor_row_factory(supplier="C")
+    results[ResultKeyName.grid_loss_sys_cor_master_data] = sys_cor_row_factory(
+        supplier="C"
+    )
 
-    result_df = adjust_flex_consumption(results, metadata)
+    result_df = adjust_production(results, metadata)
 
     result_df_collect = result_df.collect()
     assert result_df.count() == 3
@@ -344,10 +347,10 @@ def test_result_dataframe_contains_same_number_of_results_with_same_energy_suppl
     assert result_df_collect[2][Colname.energy_supplier_id] == "C"
 
 
-def test_correct_grid_loss_entry_is_used_to_determine_energy_responsible_for_the_given_time_window_from_flex_consumption_result_dataframe(
-    flex_consumption_result_row_factory,
-    added_grid_loss_result_row_factory,
-    grid_loss_sys_cor_row_factory,
+def test_correct_system_correction_entry_is_used_to_determine_energy_responsible_for_the_given_time_window_from_hourly_production_result_dataframe(
+    hourly_production_result_row_factory,
+    added_system_correction_result_row_factory,
+    sys_cor_row_factory,
 ):
     results = {}
     time_window_1 = {
@@ -363,64 +366,64 @@ def test_correct_grid_loss_entry_is_used_to_determine_energy_responsible_for_the
         Colname.end: datetime(2020, 1, 1, 3, 0),
     }
 
-    fc_row_1 = flex_consumption_result_row_factory(
+    hp_row_1 = hourly_production_result_row_factory(
         supplier="A", time_window=time_window_1
     )
-    fc_row_2 = flex_consumption_result_row_factory(
+    hp_row_2 = hourly_production_result_row_factory(
         supplier="B", time_window=time_window_2
     )
-    fc_row_3 = flex_consumption_result_row_factory(
+    hp_row_3 = hourly_production_result_row_factory(
         supplier="B", time_window=time_window_3
     )
 
     results[
-        ResultKeyName.flex_consumption
+        ResultKeyName.hourly_production
     ] = create_dataframe_from_aggregation_result_schema(
-        metadata, fc_row_1.union(fc_row_2).union(fc_row_3)
+        metadata, hp_row_1.union(hp_row_2).union(hp_row_3)
     )
 
-    gagl_result_1 = Decimal(1)
-    gagl_result_2 = Decimal(2)
-    gagl_result_3 = Decimal(3)
+    gasc_result_1 = Decimal(1)
+    gasc_result_2 = Decimal(2)
+    gasc_result_3 = Decimal(3)
 
-    gagl_row_1 = added_grid_loss_result_row_factory(
-        time_window=time_window_1, added_grid_loss=gagl_result_1
+    gasc_row_1 = added_system_correction_result_row_factory(
+        time_window=time_window_1, added_system_correction=gasc_result_1
     )
-    gagl_row_2 = added_grid_loss_result_row_factory(
-        time_window=time_window_2, added_grid_loss=gagl_result_2
+    gasc_row_2 = added_system_correction_result_row_factory(
+        time_window=time_window_2, added_system_correction=gasc_result_2
     )
-    gagl_row_3 = added_grid_loss_result_row_factory(
-        time_window=time_window_3, added_grid_loss=gagl_result_3
+    gasc_row_3 = added_system_correction_result_row_factory(
+        time_window=time_window_3, added_system_correction=gasc_result_3
     )
 
     results[
-        ResultKeyName.added_grid_loss
+        ResultKeyName.added_system_correction
     ] = create_dataframe_from_aggregation_result_schema(
-        metadata, gagl_row_1.union(gagl_row_2).union(gagl_row_3)
+        metadata, gasc_row_1.union(gasc_row_2).union(gasc_row_3)
     )
 
-    glsc_row_1 = grid_loss_sys_cor_row_factory(
+    sc_row_1 = sys_cor_row_factory(
         supplier="A", valid_from=time_window_1["start"], valid_to=time_window_1["end"]
     )
-    glsc_row_2 = grid_loss_sys_cor_row_factory(
+    sc_row_2 = sys_cor_row_factory(
         supplier="C", valid_from=time_window_2["start"], valid_to=time_window_2["end"]
     )
-    glsc_row_3 = grid_loss_sys_cor_row_factory(
+    sc_row_3 = sys_cor_row_factory(
         supplier="B", valid_from=time_window_3["start"], valid_to=None
     )
 
-    results[ResultKeyName.grid_loss_sys_cor_master_data] = glsc_row_1.union(
-        glsc_row_2
-    ).union(glsc_row_3)
+    results[ResultKeyName.grid_loss_sys_cor_master_data] = sc_row_1.union(
+        sc_row_2
+    ).union(sc_row_3)
 
-    result_df = adjust_flex_consumption(results, metadata)
+    result_df = adjust_production(results, metadata)
 
     assert result_df.count() == 3
     assert (
         result_df.filter(col(Colname.energy_supplier_id) == "A")
         .filter(col(f"{Colname.time_window_start}") == time_window_1["start"])
         .collect()[0][Colname.sum_quantity]
-        == default_sum_quantity + gagl_result_1
+        == default_sum_quantity + gasc_result_1
     )
     assert (
         result_df.filter(col(Colname.energy_supplier_id) == "B")
@@ -432,21 +435,5 @@ def test_correct_grid_loss_entry_is_used_to_determine_energy_responsible_for_the
         result_df.filter(col(Colname.energy_supplier_id) == "B")
         .filter(col(f"{Colname.time_window_start}") == time_window_3["start"])
         .collect()[0][Colname.sum_quantity]
-        == default_sum_quantity + gagl_result_3
+        == default_sum_quantity + gasc_result_3
     )
-
-    # for i in range(1, 23):
-    #     time_windows.append({'start': datetime(2020, 1, 1, i, 0), 'end': datetime(2020, 1, 1, i + 1, 0)})
-    #     sup = "A"
-    #     if i > 4:
-    #         sup = "B"
-    #     if i > 8:
-    #         sup = "C"
-    #     if i > 12:
-    #         sup = "D"
-    #     if i > 16:
-    #         sup = "E"
-    #     if i > 20:
-    #         sup = "F"
-    #     fc_row = flex_consumption_result_row_factory(supplier=sup, time_window=time_windows[i])
-    #     fc_df.union(fc_row)
