+++ source/databricks/tests/integration/calculator/steps/test_aggregate_quality.py
@@ -15,7 +15,7 @@ from datetime import datetime, timedelta
 from geh_stream.codelists import Colname
 from pyspark.sql.types import StructType, StringType, TimestampType
 from geh_stream.codelists import Quality, MarketEvaluationPointType
-from geh_stream.aggregation_utils.aggregators import aggregate_quality
+from package.steps import aggregate_quality
 import pytest
 import pandas as pd
