+++ source/databricks/tests/integration/calculator/steps/test_hourly_consumption.py
@@ -19,7 +19,7 @@ from geh_stream.codelists import (
     ResolutionDuration,
     MarketEvaluationPointType,
 )
-from geh_stream.aggregation_utils.aggregators import (
+from package.steps import (
     aggregate_hourly_settled_consumption_ga_es,
     aggregate_hourly_settled_consumption_ga_brp,
     aggregate_hourly_settled_consumption_ga,
